fix: split transport_step_yield_preview under 250-line ceiling (#4663)
Unblock Quality repo.economy_source_file_size by extracting types/next-level mapping to a sibling library.
