+++ packages/colonizethis_economy/lib/src/economy/transport_step_yield_preview.dart
@@ -10,39 +10,9 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 
 import 'tile_extraction_pipeline.dart';
 import 'tile_extraction_yield.dart';
+import 'transport_step_yield_preview_types.dart';
 
-/// Why the next transport step does or does not raise what arrives.
-enum TransportStepYieldKind {
-  raise,
-  roadPathLimit,
-  townDevelopmentLimit,
-  disconnected,
-  bindsToCapital,
-  portOnCoast,
-}
-
-/// Current vs hypothetical next effective yield after one transport work step.
-class TransportStepYieldPreview {
-  const TransportStepYieldPreview({
-    this.commodityId,
-    required this.currentEffective,
-    required this.nextEffective,
-    required this.kind,
-  });
-
-  final String? commodityId;
-  final int currentEffective;
-  final int nextEffective;
-  final TransportStepYieldKind kind;
-}
-
-/// Work-target ids matching [colonizethis_orders] (passed as strings to avoid
-/// an orders dependency in this package).
-abstract final class TransportStepWorkTargets {
-  static const buildRoad = 'build_road';
-  static const buildPort = 'build_port';
-  static const buildRail = 'build_rail';
-}
+export 'transport_step_yield_preview_types.dart';
 
 int _storedTransportLevel({
   required TileMapState tileState,
@@ -54,28 +24,6 @@ int _storedTransportLevel({
   }
   final roadLevel = tileState.roadLevel(tileKey);
   return roadLevel > 0 ? roadLevel : 0;
-}
-
-/// Next stored transport after [workTarget] on [currentTransport], or null when
-/// the step does not apply.
-int? nextStoredTransportLevel({
-  required String workTarget,
-  required int currentTransport,
-  required bool hasRoadConstructionTech,
-}) {
-  return switch (workTarget) {
-    TransportStepWorkTargets.buildRoad => switch (currentTransport) {
-      0 => 1,
-      1 when hasRoadConstructionTech => 2,
-      _ => null,
-    },
-    TransportStepWorkTargets.buildPort => currentTransport < 4 ? 4 : null,
-    TransportStepWorkTargets.buildRail => switch (currentTransport) {
-      1 || 2 => 4,
-      _ => null,
-    },
-    _ => null,
-  };
 }
 
 Map<String, int> _pathCapWithTileTransport({
@@ -130,6 +78,36 @@ int _yieldAtTransport({
   );
 }
 
+TransportStepYieldPreview? _preludeMissingPreview({
+  required String workTarget,
+  required bool connected,
+}) {
+  if (workTarget == TransportStepWorkTargets.buildPort) {
+    return const TransportStepYieldPreview(
+      kind: TransportStepYieldKind.portOnCoast,
+      currentEffective: 0,
+      nextEffective: 0,
+    );
+  }
+  if (connected &&
+      (workTarget == TransportStepWorkTargets.buildRoad ||
+          workTarget == TransportStepWorkTargets.buildRail)) {
+    return const TransportStepYieldPreview(
+      kind: TransportStepYieldKind.bindsToCapital,
+      currentEffective: 0,
+      nextEffective: 0,
+    );
+  }
+  if (!connected) {
+    return const TransportStepYieldPreview(
+      kind: TransportStepYieldKind.disconnected,
+      currentEffective: 0,
+      nextEffective: 0,
+    );
+  }
+  return null;
+}
+
 /// Hypothetical next transport step for one extractable owned tile.
 ///
 /// Returns null when [workTarget] does not bump stored transport on this tile.
@@ -176,30 +154,10 @@ TransportStepYieldPreview? computeTransportStepYieldPreview({
   );
 
   if (prelude == null) {
-    if (workTarget == TransportStepWorkTargets.buildPort) {
-      return const TransportStepYieldPreview(
-        kind: TransportStepYieldKind.portOnCoast,
-        currentEffective: 0,
-        nextEffective: 0,
-      );
-    }
-    if (connected &&
-        (workTarget == TransportStepWorkTargets.buildRoad ||
-            workTarget == TransportStepWorkTargets.buildRail)) {
-      return const TransportStepYieldPreview(
-        kind: TransportStepYieldKind.bindsToCapital,
-        currentEffective: 0,
-        nextEffective: 0,
-      );
-    }
-    if (!connected) {
-      return const TransportStepYieldPreview(
-        kind: TransportStepYieldKind.disconnected,
-        currentEffective: 0,
-        nextEffective: 0,
-      );
-    }
-    return null;
+    return _preludeMissingPreview(
+      workTarget: workTarget,
+      connected: connected,
+    );
   }
 
   final province = prelude.province;
